domainID use uint64

+++ chains/evm/voter/voter.go
@@ -196,14 +196,14 @@ func (v *EVMVoter) SubmitSignature(m *message.Message) error {
 				{"verifyingContract", "address"},
 			},
 			"PermitBridge": []core.Type{
-				{"domainID", "uint"},
+				{"domainID", "uint64"},
 				{"depositNonce", "uint64"},
 				{"resourceID", "bytes32"},
 				{"data", "bytes"}}},
 		PrimaryType: "PermitBridge",
 		Domain:      core.TypedDataDomain{Name: name, Version: version, ChainId: math.NewHexOrDecimal256(chainId.Int64()), VerifyingContract: verifyingContract.String()},
 		Message: core.TypedDataMessage{
-			"domainID":     uint(m.Source),
+			"domainID":     uint64(m.Source),
 			"depositNonce": m.DepositNonce,
 			"resourceID":   m.ResourceId,
 			"data":         m.Data,
